FIX: Mutate query response parsing fix (#85)
### Summary :memo:
At the moment the response parsing logic for mutate query is checking
whether `status` = true, however in fact all mutate query execution
return `success` in boolean instead.

Hence this is the PR which aims to fix the mist-match, and make error
message will only be printed, when the field `success` is either absent
or false.



### Test plan:

 - Make test
 - Run mutate example

---------

Signed-off-by: Andy Kwok <andy.kwok@improving.com>